refactor(frontend): share a single fileToBase64 helper
UploadModal had a robust file->base64 reader and useVisualId had its own
inline FileReader doing the same data-URL split. Extract one
fileToBase64(Blob) helper in lib/utils and use it in both. useVisualId's
fetch is already wrapped in try/catch, so the helper rejecting on a
malformed blob (instead of silently sending an empty string) is handled
gracefully.
